+++ include/LLVM_interface/ExecutionEngine.h
@@ -24,13 +24,22 @@ namespace panther{
 
 
 				template<typename T>
-				EVO_NODISCARD auto runFunction(std::string_view func_name) noexcept -> T;
+				EVO_NODISCARD auto runFunction(std::string_view func_name) noexcept -> T {
+					const uint64_t func_addr = this->getFuncAddress(func_name);
+					
+					using FuncType = T(*)(void);
+					const FuncType func = (FuncType)func_addr;
+					return func();
+				};
 
 				template<>
 				EVO_NODISCARD auto runFunction<void>(std::string_view func_name) noexcept -> void;
 
 
 				EVO_NODISCARD inline auto hasCreatedEngine() const noexcept -> bool { return this->engine != nullptr; };
+
+			private:
+				EVO_NODISCARD auto getFuncAddress(std::string_view func_name) noexcept -> uint64_t;
 		
 			private:
 				llvm::ExecutionEngine* engine = nullptr;